Student Achievement Data

The student achievement data used in Duflo, Dupas, and Kremer (2011) is available from the Harvard Dataverse (see Source below). The original dataset contains 7,022 observations and 106 variables. We extracted 5,795 observations with non-missing values for the outcome variable totalscore and retained 10 key variables.

Usage

data(ddk_2011)

Format

A data frame with 5795 rows and 10 columns.

schoolid

ID of primary school.

tracking

School sampled for tracking.

percentile

Student's percentile in initial distribution.

totalscore

Total endline score.

etpteacher

Student assigned to contract teacher.

lowstream

Student assigned to lower-ability section (if tracking school).

highstream

Student assigned to higher-ability section (if tracking school).

girl

Sex of student: female.

agetest

Age of student at time of test.

ts_std

Standardized total endline score.

Source

https://dataverse.harvard.edu/dataset.xhtml?persistentId=hdl:1902.1/16787

References

Esther Duflo, Pascaline Dupas, and Michael Kremer (2011) “Peer Effects, Teacher Incentives, and the Impact of Tracking: Evidence from a Randomized Evaluation in Kenya,” American Economic Review, 1739-1774.