Estimate the QTE under the RDD

rdq estimates QTE under the RDD with or without covariates. This function is used by rd.qte to generate QTE estimates.

Usage

rdq(y, x, d, x0, z0 = NULL, tau, h.tau, cov)

Arguments

y

a numeric vector, the outcome variable.

x

a vector (or a matrix) of covariates, the first column is the running variable.

d

a numeric vector, the treatment status.

x0

the cutoff point.

z0

the value of the covariates at which to evaluate the effects. For example, if a female dummy is included, z0 = 1 may indicate the female subgroup.

tau

a vector of quantiles of interest.

h.tau

the bandwidth values (specified for each quantile level).

cov

either 0 or 1. Set cov=1 if covariates are present in the model; otherwise set cov=0.

Value

A list with elements:

qte

QTE estimates.

qp.est

conditional quantile estimates on the right side of \(x_{0}\) (or for the D=1 group).

qm.est

conditional quantile estimates on the left side of \(x_{0}\) (or for the D=0 group).

bcoe.p

quantile regression coefficients on the right side of \(x_{0}\).

bcoe.m

quantile regression coefficients on the left side of \(x_{0}\).

Examples

# Without covariate
n = 500
x = runif(n,min=-4,max=4)
d = (x > 0)
y = x + 0.3*(x^2) - 0.1*(x^3) + 1.5*d + rnorm(n)
tlevel = seq(0.1,0.9,by=0.1)
hh = rep(2,length(tlevel))
rdq(y=y,x=x,d=d,x0=0,z0=NULL,tau=tlevel,h.tau=hh,cov=0)
#> $qte
#>            [,1]
#>  [1,] 1.8718492
#>  [2,] 1.3777156
#>  [3,] 1.2568125
#>  [4,] 0.9638824
#>  [5,] 1.3385773
#>  [6,] 1.3314579
#>  [7,] 1.4405019
#>  [8,] 1.6020048
#>  [9,] 1.8880244
#> 
#> $qp.est
#>            [,1]
#>  [1,] 0.2893188
#>  [2,] 0.4518693
#>  [3,] 0.7948673
#>  [4,] 1.0035164
#>  [5,] 1.6134203
#>  [6,] 1.6975423
#>  [7,] 2.1102606
#>  [8,] 2.4025738
#>  [9,] 3.1342573
#> 
#> $qm.est
#>              [,1]
#>  [1,] -1.58253038
#>  [2,] -0.92584633
#>  [3,] -0.46194522
#>  [4,]  0.03963407
#>  [5,]  0.27484295
#>  [6,]  0.36608441
#>  [7,]  0.66975863
#>  [8,]  0.80056899
#>  [9,]  1.24623283
#> 
#> $bcoe.p
#>            [,1]
#>  [1,] 0.2893188
#>  [2,] 0.4518693
#>  [3,] 0.7948673
#>  [4,] 1.0035164
#>  [5,] 1.6134203
#>  [6,] 1.6975423
#>  [7,] 2.1102606
#>  [8,] 2.4025738
#>  [9,] 3.1342573
#> 
#> $bcoe.m
#>              [,1]
#>  [1,] -1.58253038
#>  [2,] -0.92584633
#>  [3,] -0.46194522
#>  [4,]  0.03963407
#>  [5,]  0.27484295
#>  [6,]  0.36608441
#>  [7,]  0.66975863
#>  [8,]  0.80056899
#>  [9,]  1.24623283
#> 

# (continued) With covariates
z = sample(c(0,1),n,replace=TRUE)
y = x + 0.3*(x^2) - 0.1*(x^3) + 1.5*d + d*z + rnorm(n)
rdq(y=y,x=cbind(x,z),d=d,x0=0,z0=c(0,1),tau=tlevel,h.tau=hh,cov=1)
#> $qte
#>            [,1]     [,2]
#>  [1,] 1.0008947 2.933147
#>  [2,] 0.2872528 3.862821
#>  [3,] 0.1570105 3.525200
#>  [4,] 0.4965719 3.465202
#>  [5,] 0.6281258 2.557983
#>  [6,] 0.6596634 2.556133
#>  [7,] 0.6242624 2.932848
#>  [8,] 1.2869185 2.628948
#>  [9,] 0.5301626 2.271813
#> 
#> $qp.est
#>              [,1]     [,2]
#>  [1,] -0.33542637 1.055404
#>  [2,] -0.32268387 1.903147
#>  [3,]  0.03924549 1.970456
#>  [4,]  0.54848498 2.187264
#>  [5,]  0.88886563 2.270340
#>  [6,]  1.32382182 2.565934
#>  [7,]  1.28709106 3.295885
#>  [8,]  2.02650242 3.328979
#>  [9,]  2.05635198 3.486251
#> 
#> $qm.est
#>              [,1]         [,2]
#>  [1,] -1.33632110 -1.877742901
#>  [2,] -0.60993668 -1.959673888
#>  [3,] -0.11776497 -1.554743938
#>  [4,]  0.05191305 -1.277937879
#>  [5,]  0.26073979 -0.287643544
#>  [6,]  0.66415840  0.009800301
#>  [7,]  0.66282868  0.363037634
#>  [8,]  0.73958389  0.700030762
#>  [9,]  1.52618935  1.214437964
#> 
#> $bcoe.p
#>              [,1]     [,2]
#>  [1,] -0.33542637 1.390831
#>  [2,] -0.32268387 2.225831
#>  [3,]  0.03924549 1.931211
#>  [4,]  0.54848498 1.638779
#>  [5,]  0.88886563 1.381474
#>  [6,]  1.32382182 1.242112
#>  [7,]  1.28709106 2.008794
#>  [8,]  2.02650242 1.302477
#>  [9,]  2.05635198 1.429899
#> 
#> $bcoe.m
#>              [,1]        [,2]
#>  [1,] -1.33632110 -0.54142180
#>  [2,] -0.60993668 -1.34973720
#>  [3,] -0.11776497 -1.43697897
#>  [4,]  0.05191305 -1.32985093
#>  [5,]  0.26073979 -0.54838333
#>  [6,]  0.66415840 -0.65435809
#>  [7,]  0.66282868 -0.29979105
#>  [8,]  0.73958389 -0.03955313
#>  [9,]  1.52618935 -0.31175138
#>